dlt/requirements.txt
asn1crypto==1.5.1
boto3==1.40.28
botocore==1.40.28
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dlt==1.16.0
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.12.0
greenlet==3.2.4
hexbytes==1.3.1
humanize==4.13.0
idna==3.10
jmespath==1.0.1
jsonpath-ng==1.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
mysql-connector-python==9.4.0
orjson==3.11.3
packaging==25.0
pathvalidate==3.3.1
pendulum==3.1.0
platformdirs==4.4.0
pluggy==1.6.0
ply==3.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyodbc==5.2.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requirements-parser==0.13.0
rich==14.1.0
rich-argparse==1.7.1
s3transfer==0.14.0
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.17.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sqlglot==27.13.2
tenacity==9.1.2
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
